Move crypto utility functions to common library
